Added FastCheck Filters
 - Fast check filters allow to look for existence of an item and
prioritize false negatives instead of false positives like Bloom filters
do.
False negative here refers to the filter returning a negative, even if
it is possible that the object was previously seen, instead of false for
an object that was never seen, as in the case of bloom filters.

The code had to be adapter from
https://github.com/jmhodges/opposite_of_a_bloom_filter/blob/master/go/oppobloom/oppobloom.go
because, the existing implementation had only one method that would
check and add. We prefer the case in which can check and then add as
2 separate invocations

Signed-off-by: Ketan Umare <ketan.umare@gmail.com>